pyproject.toml
[tool.poetry]
package-mode = true
name = "lucy"
version = "4.8.7"
description = ""
license = "GPL-3.0-or-later"
authors = ["Brandon Graham Cobb <brandongrahamcobb@icloud.com>"]
readme = "README.md"
homepage = "https://brandongcobb.com"
repository = "https://github.com/brandongrahamcobb/lucy.git"
keywords = ["AI", "chemistry", "Discord", "json", "moderation", "python", "Twitch", "yaml"]
packages = [
    { include = "lucy", from = "src" },
]

[tool.poetry.dependencies]
python = "~3.12.8"
openai = "^1.58.1"
rdkit = "^2024.9.1"
aiohttp = "^3.11.11"
asyncpg = "^0.30.0"
bs4 = "^0.0.2"
discord = "^2.3.2"
opencv-python = "^4.10.0.84"
pyqt5 = "^5.15.11"
pubchempy = "^1.0.4"
pyyaml = "^6.0.2"
requests = "^2.32.3"
selenium = "^4.27.1"
quart = "^0.20.0"
torch = "^2.5.1"
twitchio = "^2.10.0"
webdriver-manager = "^4.0.2"


[build-system]
requires = ["poetry-core", "setuptools"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
mycli = "lucy:main"
